requirements.txt
feedparser>=6.0.10
tweepy>=4.14.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
Flask>=3.0.0
Flask-CORS>=4.0.0
